migrate to instrumenter API, move TracingFilter to library, rename module, other review refinements
