+++ settings.gradle.kts
@@ -278,9 +278,9 @@ include(":instrumentation:reactor-netty:reactor-netty-0.9:javaagent")
 include(":instrumentation:reactor-netty:reactor-netty-1.0:javaagent")
 include(":instrumentation:rediscala-1.8:javaagent")
 include(":instrumentation:redisson-3.0:javaagent")
-include(":instrumentation:restlet:restlet-core-1.1:javaagent")
-include(":instrumentation:restlet:restlet-core-1.1:library")
-include(":instrumentation:restlet:restlet-core-1.1:testing")
+include(":instrumentation:restlet:restlet-1.1:javaagent")
+include(":instrumentation:restlet:restlet-1.1:library")
+include(":instrumentation:restlet:restlet-1.1:testing")
 include(":instrumentation:rmi:bootstrap")
 include(":instrumentation:rmi:javaagent")
 include(":instrumentation:rocketmq-client-4.8:javaagent")
